Fix download code not knowing what the image download URL was
