+++ src/main/java/org/openstreetmap/josm/plugins/mapillary/gui/workers/MapillarySequenceDownloader.java
@@ -1,10 +1,6 @@
 // License: GPL. For details, see LICENSE file.
 package org.openstreetmap.josm.plugins.mapillary.gui.workers;
 
-import static org.openstreetmap.josm.plugins.mapillary.gui.workers.MapillaryNodesDownloader.realDownloadImages;
-
-import java.io.ByteArrayInputStream;
-import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Comparator;
@@ -16,33 +12,21 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
 
-import javax.annotation.Nonnull;
-import javax.json.Json;
-import javax.json.JsonObject;
-import javax.json.JsonReader;
-
-import org.openstreetmap.josm.data.osm.IWay;
 import org.openstreetmap.josm.data.vector.VectorDataSet;
 import org.openstreetmap.josm.data.vector.VectorNode;
-import org.openstreetmap.josm.plugins.mapillary.cache.Caches;
+import org.openstreetmap.josm.plugins.mapillary.data.mapillary.MapillaryDownloader;
 import org.openstreetmap.josm.plugins.mapillary.data.mapillary.MapillaryNode;
 import org.openstreetmap.josm.plugins.mapillary.data.mapillary.MapillarySequence;
 import org.openstreetmap.josm.plugins.mapillary.gui.layer.MapillaryLayer;
-import org.openstreetmap.josm.plugins.mapillary.spi.preferences.MapillaryConfig;
 import org.openstreetmap.josm.plugins.mapillary.utils.MapillaryImageUtils;
-import org.openstreetmap.josm.plugins.mapillary.utils.MapillarySequenceUtils;
-import org.openstreetmap.josm.plugins.mapillary.utils.api.JsonDecoder;
-import org.openstreetmap.josm.plugins.mapillary.utils.api.JsonSequencesDecoder;
 import org.openstreetmap.josm.tools.JosmRuntimeException;
 import org.openstreetmap.josm.tools.Logging;
 import org.openstreetmap.josm.tools.Pair;
-import org.openstreetmap.josm.tools.Utils;
 
 /**
  * Download sequences asynchronously
  */
-public class MapillarySequenceDownloader extends MapillaryDownloader<MapillarySequence, MapillarySequence> {
-    private static final long[] EMPTY_LONG = new long[0];
+public class MapillarySequenceDownloader extends MapillaryUIDownloader<MapillarySequence, MapillarySequence> {
     private static final byte STEP = 25;
     private static MapillarySequenceDownloader currentSequence;
 
@@ -93,7 +77,7 @@ public class MapillarySequenceDownloader extends MapillaryDownloader<MapillarySe
     @Override
     protected MapillarySequence doInBackground() {
         updateCurrentSequenceDownload(this);
-        final long[] images = downloadSequence(this.sequenceKey);
+        final long[] images = MapillaryDownloader.downloadSequence(this.sequenceKey);
         // This will be null if the image is null
         this.currentNodes = new MapillaryNode[images.length];
         int location = -1;
@@ -124,7 +108,7 @@ public class MapillarySequenceDownloader extends MapillaryDownloader<MapillarySe
     }
 
     private void getImageRange(int i, long[] imagesToGet, long[] images) {
-        Map<String, List<MapillaryNode>> map = realDownloadImages(imagesToGet);
+        Map<String, List<MapillaryNode>> map = MapillaryDownloader.downloadImages(imagesToGet);
         if (map.size() != 1 || !map.containsKey(this.sequenceKey)) {
             throw new IllegalArgumentException(
                 "Mapillary sequence " + this.sequenceKey + " did not download any images");
@@ -192,44 +176,5 @@ public class MapillarySequenceDownloader extends MapillaryDownloader<MapillarySe
         if (!chunks.isEmpty() && counter.getAndAdd(chunks.size()) < 3) {
             this.updater.accept(chunks.get(chunks.size() - 1));
         }
-    }
-
-    /**
-     * Download a specific set of sequences
-     *
-     * @param sequence The sequences to download
-     * @return The downloaded sequences
-     */
-    private static long[] downloadSequence(@Nonnull String sequence) {
-        // prevent infinite loops. See #20470.
-        if (Arrays.stream(Thread.currentThread().getStackTrace())
-            .skip(/* getStackTrace, downloadSequences(sequences), downloadSequences(force, sequences) */ 3)
-            .filter(element -> MapillaryDownloader.class.getName().equals(element.getClassName()))
-            .filter(element -> "downloadSequences".equals(element.getMethodName())).count() > 2) {
-            return EMPTY_LONG;
-        }
-        String strippedSequence = Utils.strip(sequence);
-        if (MapillaryLayer.hasInstance() && MapillaryLayer.getInstance().getImage() != null) {
-            IWay<?> way = MapillaryLayer.getInstance().getImage().getReferrers().stream().filter(IWay.class::isInstance)
-                .map(IWay.class::cast).filter(w -> strippedSequence.equals(MapillarySequenceUtils.getKey(w)))
-                .findFirst().orElse(null);
-            if (way != null && way.getNodesCount() >= 2) {
-                return way.getNodeIds().stream().mapToLong(Long::longValue).toArray();
-            }
-        }
-        final String url = MapillaryConfig.getUrls().getImagesBySequences(sequence);
-        final String response = Caches.META_DATA_CACHE.get(url, () -> {
-            final JsonObject urlResponse = getUrlResponse(url);
-            return urlResponse == null ? null : urlResponse.toString();
-        });
-        if (response != null) {
-            try (JsonReader reader = Json
-                .createReader(new ByteArrayInputStream(response.getBytes(StandardCharsets.UTF_8)))) {
-                JsonObject obj = reader.readObject();
-                return JsonDecoder.decodeData(obj, JsonSequencesDecoder::decodeSequence).stream()
-                    .mapToLong(Long::longValue).toArray();
-            }
-        }
-        return EMPTY_LONG;
     }
 }
